removed DashMap. Made performance improvements.
