+++ src/anomaly/mod.rs
@@ -1,4 +1,3 @@
 pub use chaoda::Chaoda;
 
-#[allow(clippy::module_inception)]
 mod chaoda;
