--- src/cakes/mod.rs
@@ -1,3 +1,0 @@
-pub use search::Search;
-
-mod search;
