+++ src/dataset.rs
@@ -1,11 +1,19 @@
-/// CLAM Dataset.
-///
-/// This module contains the declaration and definition of the Dataset struct.
-use std::{fmt, result};
-use std::fmt::Debug;
-use std::sync::Arc;
+//! Dataset trait and some structs implementing it.
+//!
+//! Contains the declaration and definition of the Dataset trait
+//! and the RowMajor struct implementing Dataset to serves most of the use cases for CLAM.
+//!
+//! TODO: Implement more structs for other types of datasets.
+//! For example:
+//! * FASTA/FASTQ files containing variable length genomic sequences.
+//! * Images. e.g. from SDSS-MaNGA dataset
+//! * Molecular graphs with Tanamoto distance.
 
-use dashmap::DashMap;
+use std::collections::HashMap;
+use std::fmt::Debug;
+use std::sync::{Arc, Mutex};
+use std::{fmt, result};
+
 use ndarray::prelude::*;
 use rand::seq::IteratorRandom;
 use rayon::prelude::*;
@@ -16,10 +24,15 @@ use crate::prelude::*;
 ///
 /// All datasets supplied to CLAM must implement this trait.
 ///
+/// The `cache` for a dataset is optional.
+/// It is meant to function as a memo-table and can save significant time
+/// when the `Metric` used is computationally expensive.
 pub trait Dataset<T, U>: Debug + Send + Sync {
     /// Returns the name of the metric used to compute the distance between instances.
     ///
-    /// :warning: This name must be available in the distances crate.
+    /// Warning: This name must be available in the distances crate.
+    ///
+    /// TODO: change this return type to a closure?
     fn metric(&self) -> &'static str; // should this return the function directly?
 
     /// Returns the number of instances in the dataset.
@@ -35,29 +48,29 @@ pub trait Dataset<T, U>: Debug + Send + Sync {
     ///
     /// # Arguments
     ///
-    /// * `index` - The index of the instance to return in the dataset.
+    /// * `index` - The index of the instance to return from the dataset.
     ///
     fn instance(&self, index: Index) -> Arc<ArrayView<T, IxDyn>>;
 
-    /// Selects `n` unique instances from the given indices and returns their indices.
+    /// Returns `n` unique instances from the given indices and returns their indices.
     ///
     /// # Arguments
     ///
-    /// * `indices` - The indices to select instances from
-    /// * `n` - The number of indices to return
-    ///
+    /// * `n` - The number of unique instances
+    /// * `indices`:
+    ///   - Some - Select unique n from given indices.
+    ///   - None - Select unique n from all indices.
     fn choose_unique(&self, indices: Indices, n: usize) -> Indices;
 
-    /// Computes the distance between the two instances at the indices provided.
+    /// Returns the distance between the two instances at the indices provided.
     ///
     /// # Arguments
     ///
     /// * `left` - Index of an instance to compute distance from
     /// * `right`- Index of an instance to compute distance from
-    ///
     fn distance(&self, left: Index, right: Index) -> U;
 
-    /// Computes the distances from the instances at left to right.
+    /// Returns the distances from the instance at left to all instances with indices in right.
     ///
     /// # Arguments
     ///
@@ -66,20 +79,21 @@ pub trait Dataset<T, U>: Debug + Send + Sync {
     #[allow(clippy::ptr_arg)]
     fn distances_from(&self, left: Index, right: &Indices) -> Vec<U>;
 
-    /// Computes and returns distances amongst the instances at left and right.
+    /// Returns distances from the instances with indices in left to the instances
+    /// with indices in right.
     ///
     /// # Arguments
     ///
-    /// * `left` - Reference to the indices of instances to compute distances among
-    /// * `right` - Reference to the indices of instances to compute distances among
+    /// * `left` - Indices of instances
+    /// * `right` - Indices of instances
     #[allow(clippy::ptr_arg)]
     fn distances_among(&self, left: &Indices, right: &Indices) -> Vec<Vec<U>>;
 
-    /// Computes and returns the pairwise distances between the instances at the given indices.
+    /// Returns the pairwise distances between the instances at the given indices.
     ///
     /// # Arguments
     ///
-    /// * `indices` - Reference to indices of instances to compute pairwise distances on.
+    /// * `indices` - Indices of instances among which to compute pairwise distances.
     #[allow(clippy::ptr_arg)]
     fn pairwise_distances(&self, indices: &Indices) -> Vec<Vec<U>>;
 
@@ -87,22 +101,27 @@ pub trait Dataset<T, U>: Debug + Send + Sync {
     fn clear_cache(&self) {}
 
     /// Returns the size of the cache used for the dataset.
+    ///
+    /// If no cache is used, return None.
+    /// Otherwise, return the number of entries in the cache.
     fn cache_size(&self) -> Option<usize>;
 }
 
-/// RowMajor dataset represented as a 2-dimensional array where rows are instances and columns are attributes.
-///  
-/// A wrapper around an `Array2` of data, along with a provided `metric`, to provide an interface
-/// for computing distances between points contained within the dataset.
+/// RowMajor represents a dataset stored as a 2-dimensional array
+/// where rows are instances and columns are features/attributes.
 ///
-/// The resulting structure can make use of caching techniques to prevent repeated (potentially expensive)
-/// calls to its internal distance function.
+/// A wrapper around an `ndarray::Array2` of data, along with a `Metric`,
+/// to provide an interface for computing distances between instances
+/// contained within the dataset.
+///
+/// The resulting structure can make use of caching techniques to prevent
+/// repeated (potentially expensive) calls to its internal distance function.
 pub struct RowMajor<T: Number, U: Number> {
     /// 2D array of data
     pub data: Array2<T>,
 
     // TODO: Remove the string name and rely only on a closure (Metric<T, U>).
-    /// Metric to use to compute distances (ex "euclidean")
+    /// Metric to use to compute distances (e.g. "euclidean", "manhattan", "cosine")
     pub metric: &'static str,
 
     /// Whether this dataset should use an internal cache (recommended)
@@ -112,7 +131,7 @@ pub struct RowMajor<T: Number, U: Number> {
     function: Metric<T, U>,
 
     // The internal cache.
-    cache: DashMap<(Index, Index), U>,
+    cache: Arc<Mutex<HashMap<(Index, Index), U>>>,
 }
 
 impl<T: Number, U: Number> fmt::Debug for RowMajor<T, U> {
@@ -127,17 +146,19 @@ impl<T: Number, U: Number> fmt::Debug for RowMajor<T, U> {
 
 impl<T: Number, U: Number> RowMajor<T, U> {
     /// Create a new Dataset, using the provided data and metric, optionally use a cache.
-    pub fn new(
-        data: Array2<T>,
-        metric: &'static str,
-        use_cache: bool,
-    ) -> Result<RowMajor<T, U>, String> {
+    ///
+    /// # Arguments
+    ///
+    /// * data - a 2-dimensional array.
+    /// * name - of distance-metric to use.
+    /// * use_cache - whether to use an internal cache for storing distances.
+    pub fn new(data: Array2<T>, metric: &'static str, use_cache: bool) -> Result<RowMajor<T, U>, String> {
         Ok(RowMajor {
             data,
             metric,
             use_cache,
             function: metric_new(metric)?,
-            cache: DashMap::new(),
+            cache: Arc::new(Mutex::new(HashMap::new())),
         })
     }
 }
@@ -175,9 +196,7 @@ impl<T: Number, U: Number> Dataset<T, U> for RowMajor<T, U> {
     #[allow(clippy::ptr_arg)]
     fn choose_unique(&self, indices: Indices, n: usize) -> Indices {
         // TODO: actually check for uniqueness among choices
-        indices
-            .into_iter()
-            .choose_multiple(&mut rand::thread_rng(), n)
+        indices.into_iter().choose_multiple(&mut rand::thread_rng(), n)
     }
 
     /// Compute the distance between `left` and `right`.
@@ -185,31 +204,23 @@ impl<T: Number, U: Number> Dataset<T, U> for RowMajor<T, U> {
         if left == right {
             U::zero()
         } else {
-            let key = if left < right {
-                (left, right)
-            } else {
-                (right, left)
-            };
-            if !self.cache.contains_key(&key) {
-                self.cache.insert(
-                    key,
-                    (self.function)(
-                        Arc::new(self.data.row(left).into_dyn()),
-                        Arc::new(self.data.row(right).into_dyn()),
-                    ),
+            let key = if left < right { (left, right) } else { (right, left) };
+            if !self.cache.lock().unwrap().contains_key(&key) {
+                let distance = (self.function)(
+                    Arc::new(self.data.row(left).into_dyn()),
+                    Arc::new(self.data.row(right).into_dyn()),
                 );
+                let mut cache = self.cache.lock().unwrap();
+                cache.insert(key, distance);
             }
-            *self.cache.get(&key).unwrap()
+            *self.cache.lock().unwrap().get(&key).unwrap()
         }
     }
 
     /// Compute the distances from `left` to all points in `right`.
     #[allow(clippy::ptr_arg)]
     fn distances_from(&self, left: Index, right: &Indices) -> Vec<U> {
-        right
-            .par_iter()
-            .map(|&r| self.distance(left, r))
-            .collect::<Vec<U>>()
+        right.par_iter().map(|&r| self.distance(left, r)).collect::<Vec<U>>()
     }
 
     /// Compute distances between all points in `left` and `right`.
@@ -223,17 +234,18 @@ impl<T: Number, U: Number> Dataset<T, U> for RowMajor<T, U> {
     /// Compute the pairwise distance between all points in `indices`.
     #[allow(clippy::ptr_arg)]
     fn pairwise_distances(&self, indices: &Indices) -> Vec<Vec<U>> {
+        // TODO: Optimize this to only make distance calls for lower triangular matrix
         self.distances_among(indices, indices)
     }
 
     /// Clears the internal cache.
     fn clear_cache(&self) {
-        self.cache.clear()
+        self.cache.lock().unwrap().clear()
     }
 
     /// Returns the number of elements in the internal cache.
     fn cache_size(&self) -> Option<usize> {
-        Some(self.cache.len())
+        Some(self.cache.lock().unwrap().len())
     }
 }
