+++ src/metric.rs
@@ -1,3 +1,5 @@
+//! Number trait and Metric type.
+
 use std::collections::HashSet;
 use std::fmt::{Debug, Display};
 use std::iter::Sum;
@@ -7,19 +9,9 @@ use ndarray::{ArrayView, IxDyn};
 use ndarray_npy::{ReadableElement, WritableElement};
 use num_traits::{Num, NumCast};
 
+/// Use the Number trait to compute distances among generic numeric types.
 pub trait Number:
-    Num
-    + NumCast
-    + Sum
-    + Copy
-    + Clone
-    + PartialOrd
-    + Send
-    + Sync
-    + Debug
-    + Display
-    + ReadableElement
-    + WritableElement
+    Num + NumCast + Sum + Copy + Clone + PartialOrd + Send + Sync + Debug + Display + ReadableElement + WritableElement
 {
 }
 impl Number for f32 {}
@@ -33,8 +25,27 @@ impl Number for i16 {}
 impl Number for i32 {}
 impl Number for i64 {}
 
+/// A Metric is a function that takes two instances (generic over a Number T)
+/// and produces a single non-negative Number U.
 pub type Metric<T, U> = fn(Arc<ArrayView<T, IxDyn>>, Arc<ArrayView<T, IxDyn>>) -> U;
 
+/// Returns a Metric from a given name, or an Err if the name is not found among the implemented metrics.
+///
+/// # Arguments
+///
+/// * metric: A `&str` name of a distance function.
+/// This can be one of:
+///   - "euclidean": L2-norm.
+///   - "euclideansq": Squared L2-norm.
+///   - "manhattan": L1-norm.
+///   - "cosine": Cosine distance.
+///   - "hamming": Hamming distance.
+///   - "jaccard": Jaccard distance.
+///
+/// We plan on adding the following:
+///   - "levenshtein": Edit-distance among strings (e.g. genomic/amino-acid sequences).
+///   - "wasserstein": Earth-Mover-Distance among high-dimensional probability distributions (will be usable with images)
+///   - "tanamoto": Jaccard distance between the Maximal-Common-Subgraph of two molecular structures.
 pub fn metric_new<T: Number, U: Number>(metric: &'static str) -> Result<Metric<T, U>, String> {
     match metric {
         "euclidean" => Ok(euclidean),
@@ -52,15 +63,8 @@ fn euclidean<T: Number, U: Number>(x: Arc<ArrayView<T, IxDyn>>, y: Arc<ArrayView
     U::from(d.sqrt()).unwrap()
 }
 
-fn euclideansq<T: Number, U: Number>(
-    x: Arc<ArrayView<T, IxDyn>>,
-    y: Arc<ArrayView<T, IxDyn>>,
-) -> U {
-    let d: T = x
-        .iter()
-        .zip(y.iter())
-        .map(|(&a, &b)| (a - b) * (a - b))
-        .sum();
+fn euclideansq<T: Number, U: Number>(x: Arc<ArrayView<T, IxDyn>>, y: Arc<ArrayView<T, IxDyn>>) -> U {
+    let d: T = x.iter().zip(y.iter()).map(|(&a, &b)| (a - b) * (a - b)).sum();
     U::from(d).unwrap()
 }
 
@@ -109,10 +113,7 @@ fn jaccard<T: Number, U: Number>(x: Arc<ArrayView<T, IxDyn>>, y: Arc<ArrayView<T
     }
 
     let x: HashSet<u64> = x.iter().map(|&a| NumCast::from(a).unwrap()).collect();
-    let intersect = y
-        .iter()
-        .filter(|&&b| x.contains(&NumCast::from(b).unwrap()))
-        .count();
+    let intersect = y.iter().filter(|&&b| x.contains(&NumCast::from(b).unwrap())).count();
 
     if intersect == x.len() && intersect == y.len() {
         return U::zero();
@@ -137,54 +138,36 @@ mod tests {
         let distance = metric_new("euclideansq").unwrap();
         approx_eq!(
             f64,
-            distance(
-                Arc::new(data.row(0).into_dyn()),
-                Arc::new(data.row(0).into_dyn())
-            ),
+            distance(Arc::new(data.row(0).into_dyn()), Arc::new(data.row(0).into_dyn())),
             0.
         );
         approx_eq!(
             f64,
-            distance(
-                Arc::new(data.row(0).into_dyn()),
-                Arc::new(data.row(1).into_dyn())
-            ),
+            distance(Arc::new(data.row(0).into_dyn()), Arc::new(data.row(1).into_dyn())),
             9.
         );
 
         let distance = metric_new("euclidean").unwrap();
         approx_eq!(
             f64,
-            distance(
-                Arc::new(data.row(0).into_dyn()),
-                Arc::new(data.row(0).into_dyn())
-            ),
+            distance(Arc::new(data.row(0).into_dyn()), Arc::new(data.row(0).into_dyn())),
             0.
         );
         approx_eq!(
             f64,
-            distance(
-                Arc::new(data.row(0).into_dyn()),
-                Arc::new(data.row(1).into_dyn())
-            ),
+            distance(Arc::new(data.row(0).into_dyn()), Arc::new(data.row(1).into_dyn())),
             3.
         );
 
         let distance = metric_new("manhattan").unwrap();
         approx_eq!(
             f64,
-            distance(
-                Arc::new(data.row(0).into_dyn()),
-                Arc::new(data.row(0).into_dyn())
-            ),
+            distance(Arc::new(data.row(0).into_dyn()), Arc::new(data.row(0).into_dyn())),
             0.
         );
         approx_eq!(
             f64,
-            distance(
-                Arc::new(data.row(0).into_dyn()),
-                Arc::new(data.row(1).into_dyn())
-            ),
+            distance(Arc::new(data.row(0).into_dyn()), Arc::new(data.row(1).into_dyn())),
             5.
         );
     }
