+++ src/search/mod.rs
@@ -1,0 +1,3 @@
+pub use cakes::Cakes;
+
+mod cakes;
